refactor: migra chaves secretas para variáveis de ambiente nas funções Edge
