+++ supabase/functions/.gitignore
@@ -1,0 +1,36 @@
+# Arquivos de ambiente
+.env
+.env.local
+.env.development
+.env.test
+.env.production
+
+# Arquivos de log
+*.log
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+
+# Diretórios de dependências
+node_modules/
+.npm/
+
+# Arquivos de build
+dist/
+build/
+
+# Arquivos temporários
+.tmp/
+.temp/
+tmp/
+temp/
+
+# Arquivos de IDE
+.idea/
+.vscode/
+*.sublime-project
+*.sublime-workspace
+
+# Arquivos do sistema
+.DS_Store
+Thumbs.db
